requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
qdrant-client==1.6.9
sqlalchemy==2.0.23
alembic==1.12.1
apscheduler==3.10.4
redis==5.0.1
celery==5.3.4
twilio==8.10.3
langchain==0.0.340
langchain-openai==0.0.2
python-multipart==0.0.6
cryptography==41.0.7